WiFi utils API refactoring
- Clean up configurations (Kconfig) and dependency
- Move Lwnl80211 Kconfig into WiFi Manager's
