+++ framework/src/wifi_manager/Kconfig
@@ -13,80 +13,92 @@ config WIFI_MANAGER
 if WIFI_MANAGER
 
 choice
-	prompt "WiFi library"
-	default SLSI_WPA_SUPPLICANT
-
-config SLSI_WPA_SUPPLICANT
-	bool "select wpa_supplicant"
-	select WPA_SUPPLICANT
-	select NETUTILS_WIFI
-
-	---help---
-		Select the wpa_supplicant
-
-config SELECT_PROPIETARY_SUPPLICANT
-	bool "Select external vendor supplicant"
-	---help---
-		select the third party supplicant
-
-endchoice # WiFi supplicant library
-
-choice
 	prompt "WiFi Driver"
 	default SELECT_NO_DRIVER
 
 config SELECT_NO_DRIVER
-	bool "No Wi-Fi Driver"
+	bool "Virtual Driver"
 
 config SELECT_SCSC_WLAN
 	bool "Enable SCSC Wireless Module"
-	select SCSC_WLAN
 
 config SELECT_RTK_WLAN
 	bool "Enable RTK Wireless Module"
-	select RTK_WLAN
 
 config SELECT_PROPIETARY_WLAN
 	bool "Enable vendor-specific Wireless Module"
-
 endchoice # WiFi driver choice
 
+choice
+	prompt "WiFi library"
+	default SELECT_PROPIETARY_SUPPLICANT
+
+config SLSI_WPA_SUPPLICANT
+	bool "Enable wpa_supplicant"
+	select WPA_SUPPLICANT
+	depends on SELECT_SCSC_WLAN
+	---help---
+		Select the wpa_supplicant
+
+config SELECT_PROPIETARY_SUPPLICANT
+	depends on !SELECT_SCSC_WLAN && !SELECT_NO_DRIVER
+	bool "Enable vendor-specific supplicant"
+	---help---
+		Select the third party supplicant
+
+config SELECT_NO_SUPPLICANT
+	depends on SELECT_NO_DRIVER
+	bool "Enable virtual supplicant"
+	---help---
+		Select the virtual set of drivers
+
+endchoice # WiFi supplicant library
+
 if SLSI_WPA_SUPPLICANT
-source "$EXTERNALDIR/slsi_wifi/Kconfig"
 source "$EXTERNALDIR/wpa_supplicant/Kconfig"
 endif #SLSI_WPA_SUPPLICANT
+
+menuconfig LWNL80211
+	bool "Enable light-weight netlink for 80211"
+	default n
+	depends on !SELECT_PROPIETARY_WLAN
+	---help---
+		Lwnl80211 provides a driver interface for 802.11 WiFi hardware.
+
+if LWNL80211
+source "drivers/lwnl/Kconfig"
+endif
 
 config WIFI_PROFILE_SECURESTORAGE
 	bool "Store the Wi-Fi Profile in a Secure Storage"
 	depends on SECURITY_SS
 	default n
 
-
 config WIFIMGR_SOFTAP_IFNAME
-	string "interface name of softap mode"
+	string "Interface name of softap mode"
 	default "wl1"
 
 config WIFIMGR_STA_IFNAME
-	string "interface name of station mode"
+	string "Interface name of station mode"
 	default "wl1"
 
 config WIFIMGR_DISABLE_AUTO_GET_IP
-	bool "disable auto get ip (ipv4 dhcp)"
+	bool "Disable auto get ip (ipv4 dhcp)"
 	default n
 
 config WIFIMGR_DISABLE_DHCPS
-	bool "disable auto run dhcp server(ipv4 dhcp server)"
+	bool "Disable auto run dhcp server(ipv4 dhcp server)"
 	default n
     help
-        if say yes,it allows WiFi vendor to run dhcp server. 
+        If say yes,it allows WiFi vendor to run dhcp server. 
 
 config DISABLE_EXTERNAL_AUTOCONNECT
-	bool "disable external autoconnect"
+	bool "Disable external autoconnect"
 	default n
 
 if DISABLE_EXTERNAL_AUTOCONNECT
 config WIFIMGR_INTERNAL_AUTOCONNECT
-	bool "enable internal auto(re)connect"
+	bool "Enable internal auto(re)connect"
 	default y
 
 endif #WIFIMGR_DISABLE_EXTERNAL_AUTOCONNECT
